+++ src/Koffee.Tests/MainNavTests_SuggestPaths.fs
@@ -62,7 +62,7 @@ let ``Path suggestions return expected results`` input expected =
     let model =
         { testModel with
             LocationInput = input
-            History = { testModel.History with Paths = history |> List.map createPath }
+            History = { testModel.History with Paths = history |> List.map createHistoryPath }
         }
 
     let actual = suggestPaths fs model